Make `Error` Send + Sync + 'static
When using jpeg-decoder in a multi-threaded application, or with a
library like `anyhow`, it is desirable to have the error type meet the
`Error + Send + Sync + 'static` bounds. `jpeg_decoder::Error` previously
had an `Internal(Box<dyn std::Error>)` variant that prevented this.

The only use of this variant was for errors related to sending and
receiving messages on the worker. `SendError` and `RecvError` seem to be
designed to be handled locally rather than returned with `?`. SendError
returns the object that failed to be sent on the channel, which made it
not `Sync`, and returning an internal message object is not useful to
expose to the caller of `jpeg_decoder`.

Those channel related errors can only occur if the worker thread panics
or otherwise exits prematurely. This would be a library bug, not
something caused by user input, so it seems like the kind of fault
should just be a panic rather than a returned error.